+++ structs/key_value.go
@@ -2,5 +2,5 @@ package structs
 
 type KeyValue struct {
 	Key int64 `db:"key"`
-	Value float64 `db:"value"`
+	Value float32 `db:"value"`
 }
